Quippysync can sync a schools Active Directory student accounts with the schools student management system.
Quippysync is an application that can help schools sync their students'
Active Directory accounts with their student manaagement system (SMS).
The application has 4 main function:

1) Create Accounts - If an account exists in the schools's SMS but not
in their Active Directory, the account will be created.
2) Disable Accounts - If an account exists in the school's Active
Directory but not in their SMS, the account will be disabled.
3) Enable Accounts - If a previously disabled account reappears in the
school's SMS, the account will be enabled in Active Directory.
4) Move Accounts - If an account moves to a different OU in the
school's SMS, it will be moved to the correct OU in Active Directory.

You can see a video introduction to Quippysync at the following address:

https://www.youtube.com/watch?v=yJwK53GRRxQ